Refactor to separete out classes by format
