+++ opentile/jpeg/__init__.py
@@ -1,0 +1,2 @@
+from opentile.jpeg.jpeg import Jpeg, JpegCropError
+from opentile.jpeg.turbojpeg_patch import find_turbojpeg_path
