+++ opentile/ndpi/__init__.py
@@ -1,0 +1,1 @@
+from opentile.ndpi.ndpi_tiler import NdpiTiler
