+++ opentile/ndpi/ndpi_page.py
@@ -12,274 +12,18 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-from datetime import datetime
-from functools import cached_property
-import math
 from abc import ABCMeta, abstractmethod
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
+from functools import cached_property
+from typing import Dict, List, Optional, Sequence, Tuple
 
 import numpy as np
-from tifffile.tifffile import COMPRESSION, TIFF, FileHandle, TiffFile, TiffPage
+from tifffile.tifffile import COMPRESSION, TIFF, FileHandle, TiffPage
 
-from opentile.common import OpenTilePage, Tiler
+from opentile.tiler import OpenTilePage
 from opentile.geometry import Point, Region, Size, SizeMm
 from opentile.jpeg import Jpeg, JpegCropError
-from opentile.metadata import Metadata
-
-
-def get_value_from_ndpi_comments(
-    comments: str, value_name: str, value_type: Any
-) -> Any:
-    """Read value from ndpi comment string."""
-    for line in comments.split('\n'):
-        if value_name in line:
-            value_string = line.split("=")[1]
-            return value_type(value_string)
-
-
-class NdpiCache:
-    """Cache for bytes ordered by tile position. Oldest entry is removed when
-    size of content is above set size."""
-
-    def __init__(self, size: int):
-        """Create cache for size items.
-
-        Parameters
-        ----------
-        size: int
-            Size of the cache.
-
-        """
-        self._size = size
-        self._content: Dict[Point, bytes] = {}
-        self._history: List[Point] = []
-
-    def __len__(self) -> int:
-        return len(self._history)
-
-    def __str__(self) -> str:
-        return (
-            f"{type(self).__name__} of size {len(self)} " f"and max size {self._size}"
-        )
-
-    def __repr__(self) -> str:
-        return f'{type(self).__name__}({self._size})'
-
-    def __setitem__(self, key: Point, value: bytes) -> None:
-        """Set item in cache. Remove old items if needed.
-
-        Parameters
-        ----------
-        key: Point
-            Key for item to set.
-        value: bytes:
-            Value for item to set.
-
-        """
-        self._content[key] = value
-        self._history.append(key)
-        self._remove_old()
-
-    def __getitem__(self, key: Point) -> bytes:
-        """Get item from cache.
-
-        Parameters
-        ----------
-        key: Point
-            Key for item to get.
-
-        Returns
-        ----------
-        bytes
-            Value for key.
-        """
-        return self._content[key]
-
-    def __contains__(self, key: Point) -> bool:
-        """Return true if key in cache.
-
-        Parameters
-        ----------
-        key: Point
-            Key to check for.
-
-        Returns
-        ----------
-        bool
-            True if key is in cache.
-        """
-        return key in self._content
-
-    @property
-    def size(self) -> int:
-        return self._size
-
-    def update(self, items: Dict[Point, bytes]) -> None:
-        """Update items in cache. Remove old items if needed.
-
-        Parameters
-        ----------
-        items: Dict[Point, bytes]
-            Items to update.
-
-        """
-        self._content.update(items)
-        self._history += list(items.keys())
-        self._remove_old()
-
-    def _remove_old(self) -> None:
-        """Remove old items in cache if needed."""
-        while len(self._history) > self._size:
-            key_to_remove = self._history.pop(0)
-            self._content.pop(key_to_remove)
-
-
-class NdpiTile:
-    """Defines a tile by position and coordinates and size for cropping out
-    out frame."""
-
-    def __init__(self, position: Point, tile_size: Size, frame_size: Size) -> None:
-        """Create a ndpi tile and calculate cropping parameters.
-
-        Parameters
-        ----------
-        position: Point
-            Tile position.
-        tile_size: Size
-            Tile size.
-        frame_size: Size
-            Frame size.
-
-        """
-        self._position = position
-        self._tile_size = tile_size
-        self._frame_size = frame_size
-
-        self._tiles_per_frame = Size.max(
-            self._frame_size // self._tile_size, Size(1, 1)
-        )
-        position_inside_frame: Point = (self.position * self._tile_size) % Size.max(
-            self._frame_size, self._tile_size
-        )
-        self._left = position_inside_frame.x
-        self._top = position_inside_frame.y
-        self._frame_position = (
-            self.position // self._tiles_per_frame
-        ) * self._tiles_per_frame
-
-    def __eq__(self, other: object) -> bool:
-        if isinstance(other, NdpiTile):
-            return (
-                self.position == other.position
-                and self._tile_size == other._tile_size
-                and self._frame_size == other._frame_size
-            )
-        return NotImplemented
-
-    def __repr__(self) -> str:
-        return (
-            f'{type(self).__name__}({self.position}, {self._tile_size}, '
-            f'{self._frame_size})'
-        )
-
-    def __str__(self) -> str:
-        return f'{type(self).__name__} of position {self.position}'
-
-    @property
-    def position(self) -> Point:
-        """Return position of tile."""
-        return self._position
-
-    @property
-    def frame_position(self) -> Point:
-        """Return frame position for tile."""
-        return self._frame_position
-
-    @property
-    def left(self) -> int:
-        """Return left coordinate for tile inside frame."""
-        return self._left
-
-    @property
-    def top(self) -> int:
-        """Return top coordinate for tile inside frame."""
-        return self._top
-
-    @property
-    def width(self) -> int:
-        """Return width for tile inside frame."""
-        return self._tile_size.width
-
-    @property
-    def height(self) -> int:
-        """Return height for tile inside frame."""
-        return self._tile_size.height
-
-    @property
-    def frame_size(self) -> Size:
-        """Return frame size."""
-        return self._frame_size
-
-
-class NdpiFrameJob:
-    """A list of tiles to create from a frame. Tiles need to have the same
-    frame position."""
-
-    def __init__(self, tiles: Union[NdpiTile, List[NdpiTile]]) -> None:
-        """Create a frame job from given tile(s).
-
-        Parameters
-        ----------
-        tiles: Union[NdpiTile, List[NdpiTile]]
-            Tile(s) to base the frame job on.
-
-        """
-        if isinstance(tiles, NdpiTile):
-            tiles = [tiles]
-        first_tile = tiles.pop(0)
-        self._position = first_tile.frame_position
-        self._frame_size = first_tile.frame_size
-        self._tiles: List[NdpiTile] = [first_tile]
-        for tile in tiles:
-            self.append(tile)
-
-    def __repr__(self) -> str:
-        return f'{type(self).__name__}{self.tiles})'
-
-    def __str__(self) -> str:
-        return f'{type(self).__name__} of tiles {self.tiles}'
-
-    def __eq__(self, other: object) -> bool:
-        if isinstance(other, NdpiFrameJob):
-            return self.tiles == other.tiles
-        return NotImplemented
-
-    @property
-    def position(self) -> Point:
-        """The frame position of the frame job."""
-        return self._position
-
-    @property
-    def frame_size(self) -> Size:
-        """Frame size required for reading tiles in NdpiFrameJob."""
-        return self._frame_size
-
-    @property
-    def tiles(self) -> List[NdpiTile]:
-        """Tiles in NdpiFrameJob."""
-        return self._tiles
-
-    @property
-    def crop_parameters(self) -> List[Tuple[int, int, int, int]]:
-        """Parameters for croping tiles from frame in NdpiFrameJob."""
-        return [(tile.left, tile.top, tile.width, tile.height) for tile in self._tiles]
-
-    def append(self, tile: NdpiTile) -> None:
-        """Add a tile to the tile job."""
-        if tile.frame_position != self.position:
-            raise ValueError(f'{tile} does not match {self} frame position')
-        self._tiles.append(tile)
+from opentile.ndpi.ndpi_cache import NdpiCache
+from opentile.ndpi.ndpi_tile import NdpiFrameJob, NdpiTile
 
 
 class NdpiPage(OpenTilePage):
@@ -308,9 +52,7 @@ class NdpiPage(OpenTilePage):
         try:
             # Defined in nm
             assert isinstance(page.ndpi_tags, dict)
-            self._focal_plane = (
-                page.ndpi_tags['ZOffsetFromSlideCenter'] / 1000.0
-            )
+            self._focal_plane = page.ndpi_tags["ZOffsetFromSlideCenter"] / 1000.0
         except (KeyError, AssertionError):
             self._focal_plane = 0.0
 
@@ -357,7 +99,7 @@ class NdpiPage(OpenTilePage):
             Produced tile at position.
         """
         if tile_position != (0, 0):
-            raise ValueError('Non-tiled page, expected tile_position (0, 0)')
+            raise ValueError("Non-tiled page, expected tile_position (0, 0)")
         return self._read_frame(0)
 
     def get_decoded_tile(self, tile_position: Tuple[int, int]) -> np.ndarray:
@@ -493,9 +235,9 @@ class NdpiTiledPage(NdpiPage, metaclass=ABCMeta):
 
     def __repr__(self) -> str:
         return (
-            f'{type(self).__name__}({self._page}, {self._fh}, '
-            f'{self._base_size}, {self.tile_size}, {self._jpeg}, '
-            f'{self._frame_cache.size})'
+            f"{type(self).__name__}({self._page}, {self._fh}, "
+            f"{self._base_size}, {self.tile_size}, {self._jpeg}, "
+            f"{self._frame_cache.size})"
         )
 
     @property
@@ -712,7 +454,7 @@ class NdpiOneFramePage(NdpiTiledPage):
             Frame
         """
         if position != Point(0, 0):
-            raise ValueError('Frame position not (0, 0) for one frame level.')
+            raise ValueError("Frame position not (0, 0) for one frame level.")
         frame = self._read_frame(0)
         if (
             self.image_size.width % self.mcu.width != 0
@@ -901,269 +643,3 @@ class NdpiStripedPage(NdpiTiledPage):
             Stripe index.
         """
         return position.x + position.y * self.striped_size.width
-
-
-class NdpiMetadata(Metadata):
-    def __init__(self, page: TiffPage):
-        self._tags = page.tags
-        if page.ndpi_tags is not None:
-            self._ndpi_tags = page.ndpi_tags
-        else:
-            self._ndpi_tags = {}
-
-    @cached_property
-    def magnification(self) -> Optional[float]:
-        try:
-            return float(self._ndpi_tags["Magnification"])
-        except (AttributeError, ValueError):
-            return None
-
-    @property
-    def scanner_manufacturer(self) -> Optional[str]:
-        return self._ndpi_tags.get("Make")
-
-    @property
-    def scanner_model(self) -> Optional[str]:
-        return self._ndpi_tags.get("Model")
-
-    @property
-    def scanner_software_versions(self) -> Optional[List[str]]:
-        software_version = self._ndpi_tags.get("Software")
-        if software_version is not None:
-            return [software_version]
-        return None
-
-    @property
-    def scanner_serial_number(self) -> Optional[str]:
-        return self._ndpi_tags.get("ScannerSerialNumber")
-
-    @cached_property
-    def aquisition_datetime(self) -> Optional[datetime]:
-        datetime_tag = self._tags.get("DateTime")
-        if datetime_tag is None:
-            return None
-        try:
-            return datetime.strptime(datetime_tag.value, "%Y:%m:%d %H:%M:%S")
-        except ValueError:
-            return None
-
-    @cached_property
-    def properties(self) -> Dict[str, Any]:
-        x_offset_from_slide_center = self._ndpi_tags.get("XOffsetFromSlideCenter")
-        y_offset_from_slide_center = self._ndpi_tags.get("YOffsetFromSlideCenter")
-        z_offset_from_slide_center = self._ndpi_tags.get("ZXOffsetFromSlideCenter")
-        return {
-            "x_offset_from_slide_center": x_offset_from_slide_center,
-            "y_offset_from_slide_center": y_offset_from_slide_center,
-            "z_offset_from_slide_center": z_offset_from_slide_center,
-        }
-
-
-class NdpiTiler(Tiler):
-    # The label and overview is cropped out of the macro image.Use a faked
-    # label series to avoid clashing with series in the file.
-    FAKED_LABEL_SERIES_INDEX = -1
-
-    def __init__(
-        self,
-        filepath: Union[str, Path],
-        tile_size: int,
-        turbo_path: Optional[Union[str, Path]] = None,
-        label_crop_position: float = 0.3,
-    ):
-        """Tiler for ndpi file, with functions to produce tiles of specified
-        size.
-
-        Parameters
-        ----------
-        filepath: Union[str, Path]
-            Filepath to a ndpi TiffFile.
-        tile_size: int
-            Tile size to cache and produce. Must be multiple of 8 and will be
-            adjusted to be an even multipler or divider of the smallest strip
-            width in the file.
-        turbo_path: Optional[Union[str, Path]] = None
-            Path to turbojpeg (dll or so).
-        label_crop_position: float = 0.3
-            The position (relative to the image width) to use for cropping out
-            the label and overview image from the macro image.
-
-        """
-        super().__init__(Path(filepath))
-
-        self._fh = self._tiff_file.filehandle
-        self._tile_size = Size(tile_size, tile_size)
-        self._tile_size = self._adjust_tile_size(
-            tile_size, self._get_smallest_stripe_width()
-        )
-        if self.tile_size.width % 8 != 0 or self.tile_size.height % 8 != 0:
-            raise ValueError(f'Tile size {self.tile_size} not divisable by 8')
-        self._turbo_path = turbo_path
-        self._jpeg = Jpeg(self._turbo_path)
-
-        self._level_series_index = 0
-        for series_index, series in enumerate(self.series):
-            if series.name == "Macro":
-                self._overview_series_index = series_index
-                self._label_series_index = self.FAKED_LABEL_SERIES_INDEX
-        self._pages: Dict[Tuple[int, int, int], NdpiPage] = {}
-        self._metadata = NdpiMetadata(self.base_page)
-        self._label_crop_position = label_crop_position
-
-    def __repr__(self) -> str:
-        return (
-            f'{type(self).__name__}({self._tiff_file.filename}, '
-            f'{self.tile_size.to_tuple}, '
-            f'{self._turbo_path})'
-        )
-
-    def __str__(self) -> str:
-        return f'{type(self).__name__} of Tifffile {self._tiff_file}'
-
-    @property
-    def tile_size(self) -> Size:
-        """The size of the tiles to generate."""
-        return self._tile_size
-
-    @property
-    def metadata(self) -> Metadata:
-        return self._metadata
-
-    @classmethod
-    def supported(cls, tiff_file: TiffFile) -> bool:
-        return tiff_file.is_ndpi
-
-    def get_page(self, series: int, level: int, page: int) -> NdpiPage:
-        """Return NdpiPage for series, level, page. NdpiPages holds a cache, so
-        store created pages.
-        """
-        if not (series, level, page) in self._pages:
-            if series == self._level_series_index:
-                ndpi_page = self._create_level_page(level, page)
-            elif series == self._overview_series_index:
-                ndpi_page = self._create_overview_page()
-            elif series == self._label_series_index:
-                ndpi_page = self._create_label_page()
-            else:
-                raise ValueError("Unknown series {series}.")
-            self._pages[series, level, page] = ndpi_page
-        return self._pages[series, level, page]
-
-    @staticmethod
-    def _adjust_tile_size(
-        requested_tile_width: int, smallest_stripe_width: Optional[int] = None
-    ) -> Size:
-        """Return adjusted tile size. If file contains striped pages the
-        tile size must be an n * smallest stripe width in the file, where n
-        is the closest square factor of the ratio between requested tile width
-        and smallest stripe width.
-
-        Parameters
-        ----------
-        requested_tile_width: int
-            Requested tile width.
-        smallest_stripe_width: Optional[int] = None
-            Smallest stripe width in file.
-
-        Returns
-        ----------
-        Size
-            Adjusted tile size.
-        """
-        if (
-            smallest_stripe_width is None
-            or smallest_stripe_width == requested_tile_width
-        ):
-            # No striped pages or requested is equald to smallest
-            return Size(requested_tile_width, requested_tile_width)
-
-        if requested_tile_width > smallest_stripe_width:
-            factor = requested_tile_width / smallest_stripe_width
-        else:
-            factor = smallest_stripe_width / requested_tile_width
-        # Factor should be a square number (in the series 2^n)
-        factor_2 = pow(2, round(math.log2(factor)))
-        adjusted_width = factor_2 * smallest_stripe_width
-        return Size(adjusted_width, adjusted_width)
-
-    def _get_smallest_stripe_width(self) -> Optional[int]:
-        """Return smallest stripe width in file, or None if no page in the
-        file is striped.
-
-        Returns
-        ----------
-        Optional[int]
-            The smallest stripe width in the file, or None if no page in the
-            file is striped.
-        """
-        smallest_stripe_width: Optional[int] = None
-        for page in self._tiff_file.pages:
-            assert isinstance(page, TiffPage)
-            stripe_width = page.chunks[1]
-            if page.is_tiled and (
-                smallest_stripe_width is None or smallest_stripe_width > stripe_width
-            ):
-                smallest_stripe_width = stripe_width
-        return smallest_stripe_width
-
-    def _create_level_page(
-        self,
-        level: int,
-        page: int,
-    ) -> Union[NdpiStripedPage, NdpiOneFramePage]:
-        """Create a new level page from TiffPage.
-
-        Parameters
-        ----------
-        level: int
-            Level of page.
-        page: int
-            Page to use.
-
-        Returns
-        ----------
-        NdpiPage
-            Created page.
-        """
-        tiff_page = (
-            self._tiff_file.series[self._level_series_index].levels[level].pages[page]
-        )
-        assert isinstance(tiff_page, TiffPage)
-        if tiff_page.is_tiled:  # Striped ndpi page
-            return NdpiStripedPage(
-                tiff_page, self._fh, self.base_size, self.tile_size, self._jpeg
-            )
-        # Single frame, force tiling
-        return NdpiOneFramePage(
-            tiff_page, self._fh, self.base_size, self.tile_size, self._jpeg
-        )
-
-    def _create_label_page(self) -> CroppedNdpiPage:
-        """Create a new label page from TiffPage.
-
-        Returns
-        ----------
-        CroppedNdpiPage
-            Created page.
-        """
-        assert self._overview_series_index is not None
-        tiff_page = self._tiff_file.series[self._overview_series_index].pages.pages[0]
-        assert isinstance(tiff_page, TiffPage)
-        return CroppedNdpiPage(
-            tiff_page, self._fh, self._jpeg, (0.0, self._label_crop_position)
-        )
-
-    def _create_overview_page(self) -> NdpiPage:
-        """Create a new overview page from TiffPage.
-
-        Returns
-        ----------
-        CroppedNdpiPage
-            Created page.
-        """
-        assert self._overview_series_index is not None
-        tiff_page = self._tiff_file.series[self._overview_series_index].pages.pages[0]
-        assert isinstance(tiff_page, TiffPage)
-        return CroppedNdpiPage(
-            tiff_page, self._fh, self._jpeg, (self._label_crop_position, 1.0)
-        )
