+++ opentile/philips/__init__.py
@@ -1,0 +1,1 @@
+from opentile.philips.philips_tiff_tiler import PhilipsTiffTiler
