+++ opentile/svs/__init__.py
@@ -1,0 +1,1 @@
+from opentile.svs.svs_tiler import SvsTiler
